+++ src/main/java/DomainModel/Room.java
@@ -5,11 +5,11 @@ import java.time.LocalDateTime;
 
 public abstract class Room {
     private int number;
-    private short seats; //TODO: controlla se va bene FAI COMBACIARE COL DB
+    private int seats; //TODO: controlla se va bene FAI COMBACIARE COL DB
 
     abstract public boolean isAvailable(LocalDateTime start, Duration duration);
 
-    public Room(int number, short seats) {
+    public Room(int number, int seats) {
         if (seats <= 0)
             throw new IllegalArgumentException("Seats must be positive");
         this.number = number;
@@ -20,7 +20,7 @@ public abstract class Room {
         return number;
     }
 
-    public short getSeats() {
+    public int getSeats() {
         return seats;
     }
 }
